refactor: move rate-limit utils from @osn/core to @shared/rate-limit
createRateLimiter, getClientIp, RateLimiterBackend are generic utilities
with no OSN-specific logic. Moving to @shared/ so @zap/api (and future
packages) don't need to depend on @osn/core just for rate limiting.

- New @shared/rate-limit package with source + tests
- @osn/core: rate-limit.ts becomes a thin re-export; internal imports
  unchanged (still ../lib/rate-limit which now proxies @shared/rate-limit)
- @zap/api: import moved to @shared/rate-limit, @osn/core dep removed
- Test moved from osn/core/tests/lib/ to shared/rate-limit/tests/
