+++ osn/core/package.json
@@ -18,6 +18,7 @@
     "@osn/crypto": "workspace:*",
     "@osn/db": "workspace:*",
     "@shared/observability": "workspace:*",
+    "@shared/rate-limit": "workspace:*",
     "@shared/redis": "workspace:*",
     "@simplewebauthn/server": "^13.1.1",
     "drizzle-orm": "^0.45.0",
