+++ shared/rate-limit/vitest.config.ts
@@ -1,0 +1,2 @@
+import { defineConfig } from "vitest/config";
+export default defineConfig({ test: { environment: "node", include: ["tests/**/*.test.ts"] } });
