+++ zap/api/package.json
@@ -18,8 +18,8 @@
   "dependencies": {
     "@elysiajs/cors": "^1.4.0",
     "@elysiajs/eden": "^1.2.0",
-    "@osn/core": "workspace:*",
     "@shared/observability": "workspace:*",
+    "@shared/rate-limit": "workspace:*",
     "@zap/db": "workspace:*",
     "drizzle-orm": "^0.45.0",
     "effect": "^3.19.19",
